+++ test/script/skyscrapers.js
@@ -1,0 +1,13 @@
+/* skyscrapers.js */
+
+ui.debug.addDebugData('skyscrapers', {
+	url : '4/4/k13h4j3g',
+	failcheck : [
+		['nmDupRow', "pzprv3/skyscrapers/4/4/. . . . . . /. . . . . . /4 1 2 3 3 . /. . . . . 3 /. . . . . . /. . 1 3 . . /"],
+		['nmSightNe',"pzprv3/skyscrapers/4/4/. . . . . . /. . . . . . /4 1 2 3 4 . /. 4 . . . 3 /. . 4 . . . /. . 1 3 . . /"],
+		['ceNoNum',  "pzprv3/skyscrapers/4/4/. . . . . . /. . 1 4 3 . /4 1 2 3 4 . /. 4 3 1 2 3 /. 3 4 2 1 . /. . 1 3 . . /"],
+		[null,       "pzprv3/skyscrapers/4/4/. . . . . . /. 2 1 4 3 . /4 1 2 3 4 . /. 4 3 1 2 3 /. 3 4 2 1 . /. . 1 3 . . /"]
+	],
+	inputs : [
+	]
+});